Handle thrift unions correctly when recursing
Union fields have the same `type` value as structs (`Types::STRUCT`),
but they need to be handled differently:
The field should be validated *only* if it is the one currently set in
the union. If a different field of the union is set, then the recursion
should not continue into the not-set structures.